BLD: add dependency to build this project into a standalone application
After a few tests with building this project by `electron-builder` and
`electron-packager`, none of it was successful.

`vue-cli-plugin-electron-builder` is a plugin of `vue-cli 3` for building
electron app. This plugin works well now (for further details, please check
`README.md`), but we may need to take some time to figure out why we
cannot build this project by those 2 builders mentioned above.