BLD: add config of `extraFiles` to solve the dependency issue of `proto-loader`
`proto-loader` requires to load the *.proto files defined in
`protobufjs/google/protobuf`. These files won't be added into the source
when building application with `electron-builder`. Hence, it makes
the application failed to start when `proto-loader` is imported.

To solve this problem, those *.proto files are bundled to the distribution
by setting `extraFiles` of `electronBuilder` (a plugin of vue-cli, see also
`vue.config.js`).

This might not be a good solution, so that we have to find a better one later.