MAINT: add additional npm scripts for running test with electron
In order to run application with `grpc` enabled, developers now have
to develop in the environment of electron by running npm script
`electron:serve` (browser is not support currently, this issue might
be solved in the future).

However, when we need to test this application by selenium with running
`electron:serve` (running `python ./script/run_selenium.py dev -c 73`),
there will be two electron application opened. Thereason is that there
are 3 main phases will be done when `electron:serve` is run:
  1. Dev server launch
  2. Main build
  3. Electron launch

To avoid this, a new npm script is added: `electron:headless`.
This script will do the same things as `electron:serve` without 3rd
step.