MAINT: fix the problem of rendering project list and object saved in `$store`
`Array` and `Object` in javascript will be converted as an `Observer`
(__ob__) object after they are saved in the `$store`.

That make us hard to handle the object retrieved from `$store` (because
of the additional information).

To restore them back to normal javascript object, a new helper function
`clone` is defined in `src/api/vue_utils.js`.

Besides, `ProjectOverviewPanel` works fine now.