MAINT: set `webSecurity` to false to enable loading local image
There is a technical difficulty of loading image from local. If we
want to render a image with `<img src="local_path"/>`, we might
encounter 2 problems:

  1. `src` should be a static value, which can be determined in
    compilation stage. Otherwise, image won't be loaded correctly.

    Most answered solution is that it have to replaced by
    `src="require(local_path)"`.

    But it still not work for this case. Because the `require` syntax
    will be parsed into other form by webpack. And that leads this
    solution failed.
    See also: https://webpack.js.org/guides/dependency-management/

  2. If we use the full path of local file as `src`, another error
    will be popped out: `Not allowed to load local resource: ...`

Current solution for this issue is that we disable the `webSecurity`
in electron settings.

This might not be a good solution, further study for this issue is
required.