UPDATE: implement new service `getTrainingOutput`
In order to make user able to go to evaluation phase, we have to get
the output of training (e.g. location of output folder).

This operation will be triggered automatically when
`trainingService.StartTraining` is done (event handler `finishTraining`
will be called). And the information of training output will be saved
into `$store`. Then we can retrieve it by calling
`this.$store.getters.trainingOutput`.