MAINT: modify the implementation of maximize window when application is launched
By calling `window.maximize()` after the window is instantiated, there
will be an afterimage at the original location of application.

So that we change the implementation. We retrieve the information of
user's display, and assign it to the initalization method of electron
window. So that we don't need to call `window.maximize()` after window
is instantiated. Therefore, no afterimage will be generated.