UPDATE: finish the process flow of prediction
Complete process flow of prediction:
  1. select a model
  2. import a dataset for test
  3. run prediction
  4. go to viewer to check the prediction result

Beside, there is one thing should be cared:
We have to make the event `viewerDatasetChanged` be fired in async operation.
Otherwise, the event listener in `SideBarMenuRight` won't be triggered. (this
issue is related to the order of Vue component instantiation. for more details,
please check the commit in `ViewerOverviewPanel.vue`)