MAINT: disable the `withEscape` option of `readJson` in `Tabs.vue` to false
The reason that we set this option to true in previous version is that
the json file of validation output contains escape notations.

That is because the raw content is treated as string and dumped into
a json file in backend service. That issue is solved now, so that we
don't have to set `withEscape` now.