MAINT: fix problems related to issue #11 and rewrite `dataset_service/importDataset`
Cause of issue #11 is that we passing an `File` object rather than a string to the
argument `labelFile` in `dataset_service/importDataset`. To fix it, we just have to
select the `path` attribute of `pathInfo` returned by `<b-form-file/>`.

Besides, we add the same components for user to select desired column of filename and
label in the header of given label file. And we rewrite the function signature
of `dataset_service/importDataset` to make it able take optional argument by passing
object to it.