BUG: array in vuex store is failed to be reset
Array is store as a reference in vuex store, so that we cannot reset it
by simply assign the original value to it.
To solve this issue, `store/modules/index/initializeState` is rewritten
with checking the whether iterated object is an `Array` or not. If true,
it assign a new `Array` object to it instead.