BUG: modified annotation was not shown in `LabelPanel`
Cause of this issue is that the source of `selectedLabel` of `LabelPanel` comes
from `Viewer/parsedFileList`. When we modified an annotation of file, it did
change the annotation in background, but the source of `selectedLabel` does not
changed. That why the displayed annotation will be restored to the original one
once user reviewed the image of which the annotation was modified.