MAINT: modify the method to check OS for removing folder and suppress stderr of git command
In previous version, `which('rmdir') is None` will fail when user is using
a Windows terminal (cmd.exe). Now, we use the OS platform to select the command
for removing folders.

Besides, `stderr` is now redirected to `subprocess.DEVNULL`. So that git
credential won't be displayed on terminal once user failed to login.