BUG: fix the algorithm of finding differeces between ground truth and predicted labels
In previous implement of algorithm of finding differences between ground
truth and predicted labels, the first element in given arrays will be
ignored even if it is different in both arrays.

To solve this issue, we have to set the initial value in `Array.reduce()`
to `[]`. Besides, we have to replace the first argument `accumulator` in
callback function.

Besides, to handle with labels (annotations) indicating empty value, we
add an additional condition into that callback function (check whether an
label is `#` or not).

ref: http://10.0.4.52:3000/muen/deepra/commit/ebeeb1a