MAINT: keep imported test dataset until user import another new one
In order to make user feel more convenient, we move the function call
of reseting state of imported dataset from the timing of `entering the
page of importing test dataset` to `confirming to import test dataset`.

So that user can still view the imported dataset if they were accidentally
redirected to another pages in testing progress.